Texture Scrolling Interpolation (#835)
* Texture Scrolling Interpolation

* Use floats for texture scrolling for increased precision

* Formatting

+++ src/graphic/Fast3D/gfx_pc.h
@@ -183,8 +183,8 @@ struct RDP {
         uint8_t siz;
         uint8_t cms, cmt;
         uint8_t shifts, shiftt;
-        uint16_t uls, ult, lrs, lrt; // U10.2
-        uint16_t tmem;               // 0-511, in 64-bit word units
+        float uls, ult, lrs, lrt; // U10.2
+        uint16_t tmem;            // 0-511, in 64-bit word units
         uint32_t line_size_bytes;
         uint8_t palette;
         uint8_t tmem_index; // 0 or 1 for offset 0 kB or offset 2 kB, respectively
